merging
merging item and order features
